Move Spymemcached extractor tests to getter coverage
